+++ dev/sparktestsupport/modules.py
@@ -767,8 +767,16 @@ pyspark_pandas_slow = Module(
         "pyspark.pandas.tests.groupby.test_stat",
         "pyspark.pandas.tests.test_indexing",
         "pyspark.pandas.tests.test_ops_on_diff_frames",
-        "pyspark.pandas.tests.test_ops_on_diff_frames_slow",
         "pyspark.pandas.tests.test_ops_on_diff_frames_groupby",
+        "pyspark.pandas.tests.diff_frames_ops.test_align",
+        "pyspark.pandas.tests.diff_frames_ops.test_basic_slow",
+        "pyspark.pandas.tests.diff_frames_ops.test_cov_corrwith",
+        "pyspark.pandas.tests.diff_frames_ops.test_dot_frame",
+        "pyspark.pandas.tests.diff_frames_ops.test_dot_series",
+        "pyspark.pandas.tests.diff_frames_ops.test_index",
+        "pyspark.pandas.tests.diff_frames_ops.test_series",
+        "pyspark.pandas.tests.diff_frames_ops.test_setitem_frame",
+        "pyspark.pandas.tests.diff_frames_ops.test_setitem_series",
         "pyspark.pandas.tests.series.test_all_any",
         "pyspark.pandas.tests.series.test_arg_ops",
         "pyspark.pandas.tests.series.test_as_of",
@@ -976,8 +984,16 @@ pyspark_pandas_slow_connect = Module(
         "pyspark.pandas.tests.connect.groupby.test_parity_stat",
         "pyspark.pandas.tests.connect.test_parity_indexing",
         "pyspark.pandas.tests.connect.test_parity_ops_on_diff_frames",
-        "pyspark.pandas.tests.connect.test_parity_ops_on_diff_frames_slow",
         "pyspark.pandas.tests.connect.test_parity_ops_on_diff_frames_groupby",
+        "pyspark.pandas.tests.connect.diff_frames_ops.test_parity_align",
+        "pyspark.pandas.tests.connect.diff_frames_ops.test_parity_basic_slow",
+        "pyspark.pandas.tests.connect.diff_frames_ops.test_parity_cov_corrwith",
+        "pyspark.pandas.tests.connect.diff_frames_ops.test_parity_dot_frame",
+        "pyspark.pandas.tests.connect.diff_frames_ops.test_parity_dot_series",
+        "pyspark.pandas.tests.connect.diff_frames_ops.test_parity_index",
+        "pyspark.pandas.tests.connect.diff_frames_ops.test_parity_series",
+        "pyspark.pandas.tests.connect.diff_frames_ops.test_parity_setitem_frame",
+        "pyspark.pandas.tests.connect.diff_frames_ops.test_parity_setitem_series",
         "pyspark.pandas.tests.connect.series.test_parity_all_any",
         "pyspark.pandas.tests.connect.series.test_parity_arg_ops",
         "pyspark.pandas.tests.connect.series.test_parity_as_of",
