+++ python/pyspark/pandas/tests/connect/diff_frames_ops/test_parity_setitem_series.py
@@ -16,53 +16,23 @@
 #
 import unittest
 
-from pyspark.pandas.tests.test_ops_on_diff_frames_slow import OpsOnDiffFramesEnabledSlowTestsMixin
+from pyspark.pandas.tests.diff_frames_ops.test_setitem_series import DiffFramesSetItemSeriesMixin
 from pyspark.testing.connectutils import ReusedConnectTestCase
 from pyspark.testing.pandasutils import PandasOnSparkTestUtils
 
 
-class OpsOnDiffFramesEnabledSlowParityTests(
-    OpsOnDiffFramesEnabledSlowTestsMixin, PandasOnSparkTestUtils, ReusedConnectTestCase
+class DiffFramesParitySetItemSeriesTests(
+    DiffFramesSetItemSeriesMixin, PandasOnSparkTestUtils, ReusedConnectTestCase
 ):
-    @unittest.skip(
-        "TODO(SPARK-43611): Fix unexpected `AnalysisException` from Spark Connect client."
-    )
-    def test_diff(self):
-        super().test_diff()
-
-    @unittest.skip(
-        "TODO(SPARK-43658): Fix unexpected `SparkConnectGrpcException` from Spark Connect client."
-    )
-    def test_frame_iloc_setitem(self):
-        super().test_frame_iloc_setitem()
-
-    @unittest.skip(
-        "TODO(SPARK-43611): Fix unexpected `AnalysisException` from Spark Connect client."
-    )
-    def test_rank(self):
-        super().test_rank()
-
-    @unittest.skip(
-        "TODO(SPARK-43659): Enable OpsOnDiffFramesEnabledSlowParityTests.test_series_eq."
-    )
-    def test_series_eq(self):
-        super().test_series_eq()
-
     @unittest.skip(
         "TODO(SPARK-43658): Fix unexpected `SparkConnectGrpcException` from Spark Connect client."
     )
     def test_series_iloc_setitem(self):
         super().test_series_iloc_setitem()
 
-    @unittest.skip(
-        "TODO(SPARK-43611): Fix unexpected `AnalysisException` from Spark Connect client."
-    )
-    def test_shift(self):
-        super().test_shift()
-
 
 if __name__ == "__main__":
-    from pyspark.pandas.tests.connect.test_parity_ops_on_diff_frames_slow import *  # noqa: F401
+    from pyspark.pandas.tests.connect.diff_frames_ops.test_parity_setitem_series import *  # noqa
 
     try:
         import xmlrunner  # type: ignore[import]
